fixed routing to only be in routes folder, app only runs it. fixed dulplicate layouts in pages and now only follow layout folder
