+++ src/App.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import AppRouter from "./router/AppRouter";
+import AppRouter from "./routes/AppRouter";
 
 const App = () => {
   return (
