+++ src/pages/Dashboard.jsx
@@ -1,12 +1,8 @@
-import MainLayout from "../layouts/MainLayout";
-
 const Dashboard = () => {
   return (
-    <MainLayout>
-      <div className="p-4">
-        <h1 className="text-2xl font-bold">Dashboard</h1>
-      </div>
-    </MainLayout>
+    <div className="p-4">
+      <h1 className="text-2xl font-bold">Dashboard</h1>
+    </div>
   );
 };
