+++ src/pages/Home.jsx
@@ -1,12 +1,8 @@
-import MainLayout from "../layouts/MainLayout";
-
 const Home = () => {
   return (
-    <MainLayout>
-      <div className="p-4">
-        <h1 className="text-2xl font-bold">Home Page</h1>
-      </div>
-    </MainLayout>
+    <div className="p-4">
+      <h1 className="text-2xl font-bold">Home Page</h1>
+    </div>
   );
 };
