+++ src/pages/NotFound.jsx
@@ -1,12 +1,8 @@
-import MainLayout from "../layouts/MainLayout";
-
 const NotFound = () => {
   return (
-    <MainLayout>
-      <div className="p-4 text-center">
-        <h1 className="text-2xl font-bold">404 - Page Not Found</h1>
-      </div>
-    </MainLayout>
+    <div className="p-4 text-center">
+      <h1 className="text-2xl font-bold">404 - Page Not Found</h1>
+    </div>
   );
 };
